pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "locutus_utilities"
description = "Tools that facilitate the development and maintenance of Locutus"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
]

dependencies = ["requests",
                "pandas",
                "google-cloud-firestore"
                ]

dynamic = ["version"]

[tool.setuptools.packages.find]
where = ["src"]  # list of folders that contain the packages (["."] by default)

[project.scripts]
run_ontology_api_etl = "locutus_util.ontology_api_etl:ontology_api_etl"
drop_seed_data = "locutus_util.drop_seed_data:drop_seed_data"
update_all_seed_data = "locutus_util.run_makefile:main"